+++ .kitchen.yml
@@ -1,36 +1,51 @@
 ---
 driver:
-  name: vagrant
-  customize:
-    memory: 1024
+  name: docker
 
 provisioner:
   name: ansible_playbook
   hosts: test-kitchen
-  requirements_path: requirements.yml
-  require_pip: true
+  require_chef_for_busser: false
+  require_ansible_omnibus: false
   require_ansible_repo: false
-  ansible_version:  2.4.3.0
-  ansible_verbose:  false
+  ansible_version: 2.5.6.0
+  ansible_verbose: false
   ansible_verbosity: 2
-  ansible_diff: true
   ansible_extra_flags: <%= ENV['ANSIBLE_EXTRA_FLAGS'] %>
+  requirements_path: requirements.yml
   ignore_paths_from_root:
     - .kitchen
+  ssh_known_hosts:
+    - github.com
 
 platforms:
-  - name: ubuntu-14.04
-    driver:
-      box: ubuntu/trusty64
   - name: ubuntu-16.04
-    driver:
-      box: ubuntu/xenial64
+    driver_config:
+      image: solita/ubuntu-systemd
+      platform: ubuntu
+      privileged: true
+      run_command: /sbin/init
+      provision_command:
+        - apt-get update --fix-missing
+        - apt-get install -y apt-transport-https iproute git python-pip curl
+        - pip install ansible pyopenssl
   - name: centos-7.2
-    driver:
-      box: bento/centos-7.2
+    driver_config:
+      image: centos/systemd
+      platform: centos
+      privileged: true
+      run_command: /sbin/init
+      provision_command:
+        - yum install -y iproute tuned epel-release ansible git
+        # https://github.com/CentOS/CentOS-Dockerfiles/issues/173
+        - systemctl enable systemd-user-sessions.service && ln -s /usr/lib/systemd/system/systemd-user-sessions.service /etc/systemd/system/default.target.wants/systemd-user-sessions.service
+
+verifier:
+  name: serverspec
+  default_pattern: true
 
 suites:
   - name: default
-    driver:
-      network:
-        - ["private_network", { ip: "172.29.129.177" }]
+
+transport:
+  forward_agent: true
