+++ Gemfile
@@ -1,0 +1,7 @@
+source 'https://rubygems.org'
+
+gem 'kitchen-ansible', '=0.49.1'
+gem 'kitchen-vagrant', '=1.4.0'
+gem 'kitchen-docker', '=2.9.0'
+gem 'kitchen-verifier-serverspec', '=0.7.0'
+gem 'test-kitchen', '=1.24.0'
