+++ meta/main.yml
@@ -1,7 +1,7 @@
 ---
 galaxy_info:
-  author: "Ogonna Iwunze"
-  company: None
+  author: "O2 Priority"
+  company: "O2"
   license: BSD
   min_ansible_version: 2.0
   platforms:
@@ -10,7 +10,6 @@ galaxy_info:
         - 7
     - name: Ubuntu
       versions:
-        - trusty
         - xenial
   categories:
     - telegraf
