+++ requirements.yml
@@ -1,2 +1,9 @@
-- src: wunzeco.docker
-- src: wunzeco.influxdb
+- src: git@github.com:o2-priority/ansible-docker.git
+  version: master
+  scm: git
+  name: o2-priority.docker
+
+- src: git@github.com:o2-priority/ansible-influxdb.git
+  version: master
+  scm: git
+  name: o2-priority.influxdb
