+++ tasks/install-debian.yml
@@ -16,7 +16,7 @@
 
 - name: Install telegraf apt public signing key
   apt_key:
-    file:  /var/tmp/influxdb.key
+    file: /var/tmp/influxdb.key
     state: present
 
 #- name: Download and install telegraf apt public signing key
@@ -26,11 +26,11 @@
 
 - name: Add telegraf apt repository definition to apt sources list
   apt_repository: 
-    repo:         "{{ telegraf_apt_repo }}"
+    repo: "{{ telegraf_apt_repo }}"
     update_cache: yes
-    state:        present
+    state: present
 
 - name: Install telegraf
   apt: 
-    name:  "telegraf={{ telegraf_version|default('*') }}"
+    name: "telegraf={{ telegraf_version|default('*') }}"
     state: present
