+++ tasks/install-redhat.yml
@@ -7,10 +7,10 @@
     baseurl: "{{ telegraf_yum_repo_baseurl }}"
     gpgkey: "{{ telegraf_signing_key_url }}"
     gpgcheck: yes
-    enabled:  yes
+    enabled: yes
 
 - name: Install telegraf
   yum:
-    name:  "telegraf-{{ telegraf_version|default('*') }}"
+    name: "telegraf-{{ telegraf_version|default('*') }}"
     update_cache: yes
     state: present
