+++ test/integration/default/serverspec/default_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require '/tmp/kitchen/spec/spec_helper.rb'
 
 describe package('telegraf') do
   it { should be_installed }
